Move damage logic to DamageAuthoritySystem
